klte: use HLTE KK libsec-ril.so and libqmiservices.so
* My guess here is our device has that feature where you can share data
and wifi. I think we need to add an dummy function stub to the ril class
so the option is always disabled. Until that is resloved this seems like
the best approach. I don't see any issue in logs or using the device.

If I should revert to the whole HLTE ril stack please let me know

Change-Id: Ie8b5f055a96d963b8e1c8fca7c40198883fcceeb